Refactored the api decorators
